% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\name{assumptions_chisq}
\alias{assumptions_chisq}
\title{Assumptions for chi square}
\usage{
assumptions_chisq(mydata)
}
\arguments{
\item{mydata}{a dataframe}
}
\value{

}
\description{
Assumptions for chi square
}
