% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModel.R
\name{assumptions_lm}
\alias{assumptions_lm}
\title{Assumptions for linear model}
\usage{
assumptions_lm(mydata)
}
\arguments{
\item{mydata}{a dataframe}
}
\value{

}
\description{
Assumptions for linear model
}
