% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part2Project.R
\name{chi_squared_test_gender_activity}
\alias{chi_squared_test_gender_activity}
\title{Display relevant table for chi-squared test and perform chi-square test}
\usage{
chi_squared_test_gender_activity(data_file)
}
\arguments{
\item{data_file}{a character string of the name of the raw file containing the data that is going to be tested.}
}
\value{
result of the test
}
\description{
Display relevant table for chi-squared test and perform chi-square test
}
\examples{
chi_squared_test_gender_activity("project.csv")
}
