% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\name{conclusion_chisq}
\alias{conclusion_chisq}
\title{Conclusion for Chi Squared Test}
\usage{
conclusion_chisq(myfit)
}
\arguments{
\item{myfit}{a Chi Squared Test}
}
\value{

}
\description{
Conclusion for Chi Squared Test
}
