% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModel.R
\name{conclusion_mylm}
\alias{conclusion_mylm}
\title{Conclusion for linear model}
\usage{
conclusion_mylm(myfit)
}
\arguments{
\item{myfit}{a linear model}
}
\value{

}
\description{
Conclusion for linear model
}
