% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\name{decision_chisq}
\alias{decision_chisq}
\title{Decision for Chi Squared Test}
\usage{
decision_chisq(myfit)
}
\arguments{
\item{myfit}{a Chi Squared Test}
}
\value{

}
\description{
Decision for Chi Squared Test
}
