% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquare.R
\name{hypothesis_chisq}
\alias{hypothesis_chisq}
\title{Hypothesis for chi squared}
\usage{
hypothesis_chisq(mydata)
}
\arguments{
\item{mydata}{a dataframe}
}
\value{

}
\description{
Hypothesis for chi squared
}
