% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part2Project.R
\name{linear_regression_and_hypotheses}
\alias{linear_regression_and_hypotheses}
\title{Checking Assumptions and Display Relevant Graphs for Linear Regression}
\usage{
linear_regression_and_hypotheses(file_name, x_var, y_var)
}
\arguments{
\item{file_name}{a character string of the name of the raw file containing the data that is going to be tested.}

\item{x_var}{column name in the file containing the data of the independent variable}

\item{y_var}{column name in the file containing the data of the dependent variable}
}
\value{
assumption confirmation and result of the test
}
\description{
Checking Assumptions and Display Relevant Graphs for Linear Regression
}
\examples{
linear_regression_and_hypotheses("project.csv","height","weight")
}
