% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{data}
\name{project}
\alias{project}
\title{Height, weight and physical activity of an individual}
\format{
A data set with 1000 rows and 5 columns:
\describe{
\item{ID}{Unique identification, categorica}
\item{gender}{gender of an individual, categorical}
\item{height}{height of an individual, numerical}
\item{weight}{weight of an individual, numerical}
\item{phys}{the amount of physical acitivty conducted by an individual, categorical}
...
}
}
\usage{
project
}
\description{
Height, weight and physical activity of an individual
}
\keyword{datasets}
