% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part2Project.R
\name{t_test_mean_height}
\alias{t_test_mean_height}
\title{Display t-test result between mean and height}
\usage{
t_test_mean_height(data_file)
}
\arguments{
\item{data_file}{a dataset that contains the information to be tested}
}
\value{
t-test result
}
\description{
Display t-test result between mean and height
}
\examples{
t_test_mean_height("project.csv")
}
